[#105] Refactoring 1) more suitable file names 2) moved files around
