+++ src/ml_service/.gitignore
@@ -7,8 +7,9 @@ coverage.xml
 ./**/*.iml
 .idea/*
 *.bin
-/feature_extraction/Preprocessing/train.py
-/feature_extraction/Preprocessing/Sam_Parser/NerTraining.py
-/feature_extraction/Preprocessing/Sam_Parser/french_data/*
 /feature_extraction/Clustering/HDBSCAN/cluster_dir/*
 /feature_extraction/Clustering/HDBSCAN/metrics.txt
+/feature_extraction/Preprocessing/Arek_Parser/error_words.txt
+/feature_extraction/Clustering/dbscan/cache.pickle
+/feature_extraction/Clustering/dbscan/fact_clustering.log
+/feature_extraction/Clustering/dbscan/cluster_dir/*
