+++ src/ml_service/GlobalVariables/GlobalVariable.py
@@ -4,6 +4,7 @@
 # Change these variables to make code work on your machine
 # tmp solution since we can't download models online via
 # HTTP
+
+
 class Global:
     Precedence_Directory = r"/home/charmander/Data/text_bk/"
-
