+++ src/ml_service/feature_extraction/Clustering/dbscan/dbscan.py
@@ -2,7 +2,6 @@
 from sklearn.cluster import DBSCAN
 import os
 import numpy as np
-import time
 import logging
 logger = logging.getLogger('fact_clustering')
 
@@ -24,11 +23,10 @@ def clusterFacts(factDict):
 
 
 def writeFactsToFile(factDict, labels):
-    outputDirPath = os.path.join(os.path.dirname(
-        os.path.realpath(__file__)), '../../../outputs/')
+    outputDirPath = r'cluster_dir'
     for i, sent in enumerate(factDict.keys()):
         filePath = os.path.join(outputDirPath, str(labels[i]))
         normalizedFilePath = os.path.normpath(filePath)
-        f = open(normalizedFilePath, 'a', encoding="ISO-8859-1")
+        f = open(normalizedFilePath + '.txt', 'a')
         f.write('{:.140}'.format(sent) + '\n')
         f.close()
