+++ src/ml_service/feature_extraction/Preprocessing/Arek_Parser/related_word_fetcher.py
